Dynamic index pade

+++ public/js/stock.js
@@ -7,7 +7,7 @@ function GetItems()
 {
 	$.ajax({
 		url: '/index/getItems/',
-		method: 'GET',
+		method: 'POST',
 		success: function(json)
 		{
 			var result = $.parseJSON(json),
@@ -15,23 +15,41 @@ function GetItems()
 			if (result.success == true)
 			{
 				$('tbody').html('');
-				$.each(result.items, function(index, val)
-				{
-					template = '<tr>';
-					template += '<th scope="row" align="center">' + val.id + '</th>'
-					template += '<td align="left">' + val.name + '</td>'
-					template += '<td align="center">' + val.opt + '₽</td>'
-					template += '<td align="center">' + val.rozn + '₽</td>'
-					template += '<td align="center">' + val.col + '</td>'
-					template += '<td align="center">' + val.date + '</td>'
-					template += '<td style="padding: 8px;">'
-					template += '<button class="btn btn-light btn-sm" onclick="DeleteItem(' + val.id + ')">'
-					template += '<i class="fas fa-trash-alt"></i>'
-					template += '</button>'
-					template += '</td>'
-					template += '</tr>'
-					$('tbody').append(template);
-				});
+				$('#db').html('');
+				if (result.admin == 1)
+					$.each(result.items, function(index, val)
+					{
+						template = '<tr>';
+						template += '<th scope="row" align="center">' + val.id + '</th>'
+						template += '<td align="left">' + val.name + '</td>'
+						template += '<td align="center">' + val.opt + '₽</td>'
+						template += '<td align="center">' + val.rozn + '₽</td>'
+						template += '<td align="center">' + val.col + '</td>'
+						template += '<td align="center">' + val.date + '</td>'
+						template += '<td style="padding: 8px;">'
+						template += '<button class="btn btn-light btn-sm" onclick="DeleteItem(' + val.id + ')">'
+						template += '<i class="fas fa-trash-alt"></i>'
+						template += '</button>'
+						template += '</td>'
+						template += '</tr>'
+						$('tbody').append(template);
+						template = '<option value="' + val.name + '"></option>'
+						$('#db').append(template);
+					});
+				else
+					$.each(result.items, function(index, val)
+					{
+						template = '<tr>';
+						template += '<th scope="row" align="center">' + val.id + '</th>'
+						template += '<td align="left">' + val.name + '</td>'
+						template += '<td align="center">' + val.rozn + '₽</td>'
+						template += '<td align="center">' + val.col + '</td>'
+						template += '<td align="center">' + val.date + '</td>'
+						template += '</tr>'
+						$('tbody').append(template);
+						template = '<option value="' + val.name + '"></option>'
+						$('#db').append(template);
+					});
 			}
 		}
 	});
@@ -54,10 +72,11 @@ function DeleteItem(id)
 	});
 }
 
-function Search() {
+function Search()
+{
 	$.ajax({
 		url: '/index/search/',
-		method: 'POST',
+		method: 'GET',
 		data: $('#searchForm').serializeArray(),
 		success: function(json)
 		{
@@ -66,23 +85,53 @@ function Search() {
 			if (result.success == true)
 			{
 				$('tbody').html('');
-				$.each(result.items, function(index, val)
-				{
-					template = '<tr>';
-					template += '<th scope="row" align="center">' + val.id + '</th>'
-					template += '<td align="left">' + val.name + '</td>'
-					template += '<td align="center">' + val.opt + '₽</td>'
-					template += '<td align="center">' + val.rozn + '₽</td>'
-					template += '<td align="center">' + val.col + '</td>'
-					template += '<td align="center">' + val.date + '</td>'
-					template += '<td style="padding: 8px;">'
-					template += '<button class="btn btn-light btn-sm" onclick="DeleteItem(' + val.id + ')">'
-					template += '<i class="fas fa-trash-alt"></i>'
-					template += '</button>'
-					template += '</td>'
-					template += '</tr>'
-					$('tbody').append(template);
-				});
+				if (result.admin == 1)
+					$.each(result.items, function(index, val)
+					{
+						template = '<tr>';
+						template += '<th scope="row" align="center">' + val.id + '</th>'
+						template += '<td align="left">' + val.name + '</td>'
+						template += '<td align="center">' + val.opt + '₽</td>'
+						template += '<td align="center">' + val.rozn + '₽</td>'
+						template += '<td align="center">' + val.col + '</td>'
+						template += '<td align="center">' + val.date + '</td>'
+						template += '<td style="padding: 8px;">'
+						template += '<button class="btn btn-light btn-sm" onclick="DeleteItem(' + val.id + ')">'
+						template += '<i class="fas fa-trash-alt"></i>'
+						template += '</button>'
+						template += '</td>'
+						template += '</tr>'
+						$('tbody').append(template);
+					});
+				else
+					$.each(result.items, function(index, val)
+					{
+						template = '<tr>';
+						template += '<th scope="row" align="center">' + val.id + '</th>'
+						template += '<td align="left">' + val.name + '</td>'
+						template += '<td align="center">' + val.rozn + '₽</td>'
+						template += '<td align="center">' + val.col + '</td>'
+						template += '<td align="center">' + val.date + '</td>'
+						template += '</tr>'
+						$('tbody').append(template);
+					});
+			}
+		}
+	});
+}
+
+function Arrival()
+{
+	$.ajax({
+		url: '/index/arrival/',
+		method: 'POST',
+		data: $('#arrivalForm').serializeArray(),
+		success: function(json)
+		{
+			var result = $.parseJSON(json);
+			if (result.success == true)
+			{
+				GetItems();
 			}
 		}
 	});
